+++ public/js/app.js
@@ -1,10 +1,10 @@
-$(".openAttended").on("click", function () {
+$(".openAttended").on("click", function() {
   $(".userHoursSpace").hide();
   $(".userProjectsAttended").show();
   $(".userProjectsPending").hide();
   console.log("hey");
 });
-$(".openProjects").on("click", function () {
+$(".openProjects").on("click", function() {
   $(".userHoursSpace").hide();
   $(".userProjectsAttended").hide();
   $(".userProjectsPending").show();
@@ -16,7 +16,7 @@ if (projects.length === 0) {
     "<h2>Your volunteer box is currently empty. Take a peek at the latest projects posted:   <a href='/projects'>HERE<a><h2>";
   noProject.append(words);
   $(".userProjectsPending").append(noProject);
-} else { }
+}
 // $.get("/profile").then(function(data) {
 //   console.log("Hello");
 //   //console.log(data);
