+++ server.js
@@ -33,11 +33,11 @@ app.use(passport.initialize());
 app.use(passport.session());
 
 // Get Routes
-const apiRoute = require("./routes/apiRoutes");
+const apiLoginRoute = require("./routes/apiLoginRoutes");
 const htmlRoute = require("./routes/htmlRoutes");
-const userRoute = require("./routes/userRoutes");
+const userRoute = require("./routes/apiUserRoutes");
 require("./routes/apiProjectRoutes")(app);
-app.use("/api", apiRoute, userRoute);
+app.use("/api", apiLoginRoute, userRoute);
 app.use("/", htmlRoute);
 
 // logout route defined by passport
